Fix storing bug
This bug would sometimes appear in the revaultd functional tests.
The problem was that we would check if the signature was in the db,
then if it wasn't there, try to insert it. The problem with this is
that between the check and the insertion someone else might have
inserted the signature.
Here we remove the second check altoghether, and recognize if there
were any duplicates from the eventual error at insertion.